overlay: merge into libtrx
