+++ src/libtrx/include/libtrx/game/game_flow.h
@@ -6,3 +6,4 @@
 #include "./game_flow/reader.h"
 #include "./game_flow/sequencer.h"
 #include "./game_flow/types.h"
+#include "./game_flow/vars.h"
