+++ src/libtrx/include/libtrx/game/objects/vars.h
@@ -20,6 +20,7 @@ extern const GAME_OBJECT_ID g_InvObjects[];
 extern const GAME_OBJECT_ID g_WaterSpriteObjects[];
 extern const GAME_OBJECT_ID g_BossObjects[];
 extern const GAME_OBJECT_ID g_PlaceholderObjects[];
+extern const GAME_OBJECT_ID g_SecretObjects[];
 
 extern const GAME_OBJECT_PAIR g_GunAmmoObjectMap[];
 extern const GAME_OBJECT_PAIR g_ItemToInvObjectMap[];
