+++ src/tr1/game/demo.c
@@ -9,7 +9,6 @@
 #include "game/lara.h"
 #include "game/level.h"
 #include "game/output.h"
-#include "game/overlay.h"
 #include "game/random.h"
 #include "game/savegame.h"
 #include "game/shell.h"
@@ -21,6 +20,7 @@
 #include <libtrx/game/camera.h>
 #include <libtrx/game/interpolation.h>
 #include <libtrx/game/lara.h>
+#include <libtrx/game/overlay.h>
 #include <libtrx/game/phase.h>
 #include <libtrx/log.h>
