+++ src/tr1/game/level.c
@@ -7,7 +7,6 @@
 #include "game/objects/creatures/mutant.h"
 #include "game/objects/creatures/pierre.h"
 #include "game/output.h"
-#include "game/overlay.h"
 #include "game/random.h"
 #include "game/savegame.h"
 #include "game/shell.h"
@@ -29,6 +28,7 @@
 #include <libtrx/game/music.h>
 #include <libtrx/game/objects/traps/movable_block.h>
 #include <libtrx/game/option.h>
+#include <libtrx/game/overlay.h>
 #include <libtrx/log.h>
 #include <libtrx/memory.h>
 #include <libtrx/utils.h>
