+++ src/tr1/meson.build
@@ -216,7 +216,6 @@ sources = [
   'game/output/common.c',
   'game/output/draw.c',
   'game/output/func.c',
-  'game/overlay.c',
   'game/room_draw.c',
   'game/savegame/savegame.c',
   'game/savegame/savegame_bson.c',
