+++ src/tr2/game/objects/vars.h
@@ -2,6 +2,5 @@
 
 #include <libtrx/game/objects/vars.h>
 
-extern const GAME_OBJECT_ID g_SecretObjects[];
 extern const GAME_OBJECT_ID g_ShatterableObjects[];
 extern const GAME_OBJECT_ID g_SmashableObjects[];
