--- src/tr2/game/overlay.h
@@ -1,9 +1,0 @@
-#pragma once
-
-#include "global/types.h"
-
-#include <libtrx/game/overlay.h>
-
-void Overlay_DrawGameInfo(void);
-
-void Overlay_Animate(int32_t frames);
